Move calc_tct_recon_params() to utilities.py
